requirements.txt
django 
djangorestframework
django-allauth
python-dotenv
python-magic
django-storages
boto3
django-cors-headers
tensorflow
numpy
librosa
keras
psycopg-binary
